Fixed delete account. Profile pic and user settings are deleted on account deletion. Closes #37

+++ Commands/SettingsCommands/DeleteAccountCommand.cs
@@ -1,5 +1,6 @@
 ﻿using Data_Logger_1._3.Services;
 using MVVMEssentials.Commands;
+using System.IO;
 using System.Windows;
 
 namespace Data_Logger_1._3.Commands.SettingsCommands
@@ -8,12 +9,14 @@ namespace Data_Logger_1._3.Commands.SettingsCommands
     {
         private readonly AuthService _authService;
         private readonly IDataService _dataService;
+        private readonly SettingsService _settingsService;
 
 
-        public DeleteAccountCommand(AuthService authService, IDataService dataService)
+        public DeleteAccountCommand(AuthService authService, IDataService dataService, SettingsService settingsService)
         {
             _authService = authService;
             _dataService = dataService;
+            _settingsService = settingsService;
         }
 
         protected override async Task ExecuteAsync(object parameter)
@@ -25,6 +28,14 @@ namespace Data_Logger_1._3.Commands.SettingsCommands
 
                 if (result == MessageBoxResult.Yes)
                 {
+                    if (_authService.Account == null)
+                        throw new InvalidOperationException("Account cant be null.");
+
+                    if(File.Exists(_authService.Account.ProfilePic))
+                        File.Delete(_authService.Account.ProfilePic);
+
+                    _settingsService.Delete(_authService.Account.accountID);
+
                     var accountDeleted = await _authService.DeleteAccountAsync();
 
                     if(accountDeleted)
